Fixing up calling and part.Resource adding and removing
Still need to implement the massAdding future, right now tanks just get
really light when the fuel is removed.
One other probem. If there is no fuel in the tank, it cannot be refueled
becasue as soon as you try to activate the black hole, it turns itself
off again becasue it is empty. Should we add a "Refuel" button?